+++ src/ByteDev.Configuration.Environment/IEnvironmentVariableProvider.cs
@@ -13,6 +13,15 @@ namespace ByteDev.Configuration.Environment
         void Delete(string name);
 
         /// <summary>
+        /// Delete an environment variable. If the variable does not exist then an exception
+        /// of type <see cref="T:ByteDev.Configuration.Environment.EnvironmentVariableNotExistException" /> is thrown.
+        /// </summary>
+        /// <param name="name">Name of environment variable.</param>
+        /// <exception cref="T:System.ArgumentException"><paramref name="name" /> was null or empty.</exception>
+        /// <exception cref="T:ByteDev.Configuration.Environment.EnvironmentVariableNotExistException">Environment variable does not exist.</exception>
+        void DeleteOrThrow(string name);
+
+        /// <summary>
         /// Determines if an environment variable exists.
         /// </summary>
         /// <param name="name">Name of environment variable.</param>